At this point, the game can ask for and output daily temperature, weekly temperature, total inventory, looking at the store and then purchasing from the store. Once they have purchased from the store, their new inventory posts along with their total cash left. Next will be to give the player a chance to make their own recipe

+++ LemonadeStand/Player.cs
@@ -8,7 +8,7 @@ namespace LemonadeStand
     {
         public Inventory MyInventory;
         public Recipe MyRecipe;
-        public int MyMoney;
+        public double MyMoney;
         public double MyPrice;
         public string MyName;
